feat: Implement delivery, route, and warehouse management entities and services
- Added Delivery entity with fields for delivery details and status management.
- Created Route entity to manage routes, including vehicle and driver associations.
- Introduced RouteDelivery entity to link deliveries with routes and manage stop sequences.
- Developed Warehouse entity to store warehouse information and manage capacity.
- Implemented repositories for Delivery, Route, RouteDelivery, and Warehouse entities.
- Created services for Delivery, RouteDelivery, Route, and Warehouse with CRUD operations.
- Added pre-persist logic to set default values for status and createdAt fields.
- Established DTOs for request and response handling in service methods.

+++ backend/src/main/java/com/optimizers/backend/controller/VehicleController.java
@@ -40,21 +40,21 @@ public class VehicleController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<VehicleResponseDTO> getVehicleById(@PathVariable Long id) {
+    public ResponseEntity<VehicleResponseDTO> getVehicleById(@PathVariable Integer id) {
         VehicleResponseDTO response = vehicleService.getVehicleById(id);
         return ResponseEntity.ok(response);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<VehicleResponseDTO> updateVehicle(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @RequestBody VehicleRequestDTO requestDTO) {
         VehicleResponseDTO response = vehicleService.updateVehicle(id, requestDTO);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteVehicle(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteVehicle(@PathVariable Integer id) {
         vehicleService.deleteVehicle(id);
         return ResponseEntity.noContent().build();
     }
